Add a Continuous Deployment workflow
This requires an environment of the name `deploy-to-azure` with the
secret `AZURE_FUNCTIONAPP_PUBLISH_PROFILE` to be available. This secret
needs to contain the contents of the "publish profile" that can be
downladed in the Azure Function blade in the `Overview` section via the
`Get publish profile` link.

Since it only makes sense to deploy the App from the primary repository
(and not from any forks), we skip the deployment in forks.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>