Validate the webhook signature
Thanks to a secret stored with the GitHub App, there is a relatively
easy way to validate that GitHub sent the payload.

Let's do that, and refuse to do _anything_ if there is no signature or
the signature is invalid.

For this to work, the `GITHUB_WEBHOOK_SECRET` variable has to be set in
the Azure Function's `Application Settings` tab of the `Configuration`
section.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>