Delete installation as soon as a new one is reported
To do this, we need to use the app authentication, which in turn
requires the private key of the GitHub App. The private key has an
idiomatic, multi-line format, and won't work as a single line. However,
the only way to store this in a Function App is as an environment
variable, i.e. single line. To help with this, we store the inner part
(which can lose its newline characters and still work) under the key
`GITHUB_APP_PRIVATE_KEY` in the `Application settings` tab of the
`Configuration` section in the Function App blade.

The strategy implemented in this commit uses a hand-rolled JWT-signing
function because we do not need to pull in ~600kB worth of Javascript
_just_ for the `jsonwebtoken` package. The JWT signing is actually
pretty trivial, once you know how it works.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>