https-request: collect the response body more efficiently
String concatenation is a pretty hefty business. Better accumulate the
chunks as buffers (which requires us to drop the `setEncoding('utf-8')`
call, otherwise we receive strings) in an array, and at the end convert
that in one fell swoop.

While at it, also provide better error reporting by including the body
of the response.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>