https-request: handle 204 gracefully
An HTTP code 204 is received e.g. when sending a `workflow_dispatch`.
While we do not do that in this here GitHub App, we want to be able to
develop derivative GitHub Apps that might do that. So let's support it
by returning something more helpful than an empty string in that case.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>